requirements.txt
# THIS IS AN AUTO-GENERATED FILE. DO NOT EDIT!
# production requirements
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pytconf
pylogconf
# dev requirements
pyclassifiers
pypitools
pydmt
Sphinx
black
# test requirements
pylint
pyflakes
flake8
pytest
